change to the directory structure of components
